+++ project/app/Models/Kabupaten.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Kabupaten extends Model
+{
+    use HasFactory;
+}
