+++ project/routes/web.php
@@ -4,7 +4,8 @@ use App\Http\Controllers\Admin;
 use App\Models\Provinsi;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\HomeController;
-use App\Http\Controllers\ProvinsiController;
+use App\Http\Controllers\Admin\ProvinsiController;
+use App\Http\Controllers\Admin\KabupatenController;
 use App\Http\Controllers\Admin\RolesController;
 use App\Http\Controllers\Admin\UsersController;
 use App\Http\Controllers\Admin\CountryCountroller;
